pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["RaghadAbdulhadi <raghadabdulhadi1999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
